% Generated by roxygen2 (4.0.0): do not edit by hand
\name{combine.results.random.jungle.runs}
\alias{combine.results.random.jungle.runs}
\title{Combine results of random jungle runs.}
\usage{
combine.results.random.jungle.runs(res.prefix, no.runs, map.file = NULL,
  out.prefix)
}
\arguments{
\item{res.prefix}{a character string with the prefix of the output files of Random Jungle runs ('run_<x>' will be added automatically)}

\item{no.runs}{number of forests that should be combined (function will combined forests 1, 2, ..., <no.runs>}

\item{map.file}{a character string with name of map file (either MAP or BIM format)}

\item{out.prefix}{a character string with the prefix of the output files. The function generates two files: <out.prefix>_all_runs.info (errors and minimal importance scores) and <out.prefix>_all_runs_relative.importance}
}
\description{
This function combines variable importance scores and errors of several random jungle runs. It calculates relative importance scores for each variable in each run as well as minimal and median relative importance.
}

