% Generated by roxygen2 (4.0.0): do not edit by hand
\name{manhattanplot}
\alias{manhattanplot}
\title{Manhattan plot.}
\usage{
manhattanplot(res = NULL, results.file = NULL, rel.imp = TRUE,
  col = NULL, var.highlight = NULL, col.highlight = "black",
  threshold = NULL)
}
\arguments{
\item{res}{data.frame with results or NULL if results should be loaded from results.file}

\item{results.file}{a character string with the name of the results file or NULL if parameter res is used}

\item{rel.imp}{logical. If 'TRUE' function assumes that results from random forest variable selection should be plotted (columns 'varname', 'chr', 'pos', 'rel_imp_min'). If 'FALSE' function assumes that results of statistical test should be plotted (columns 'SNP', 'CHR', 'BP', 'P').}

\item{col}{vector of colors (if the length is smaller than the number of chromosomes colors will be recycled)}

\item{var.highlight}{vector with SNPs that will be plotted in red}

\item{col.highlight}{color for SNPs given in var.highlight (default: black)}

\item{threshold}{vector giving values at which a horizontal line should be drawn (Note: given as untransformed threshold for statistical test)}
}
\description{
This function generates a Manhattan plot for minimal relative importance scores of random forest variable selection or log transformed p-values of a statistical test.
}

