% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prepare.input.file.rj}
\alias{prepare.input.file.rj}
\title{Prepare input files for Random Jungle software.}
\usage{
prepare.input.file.rj(plink.prefix = NULL, bim.file = NULL,
  bed.file = NULL, fam.file = NULL, out.prefix, no.web = TRUE,
  recode.pheno = TRUE, zip = TRUE)
}
\arguments{
\item{plink.prefix}{character string with the prefix of binary PLINK files or NULL if bim, bed and fam files are specified separately}

\item{bim.file}{a character string with with the bim file name or NULL if plink.prefix is given}

\item{fam.file}{a character string with with the fam file name or NULL if plink.prefix is given}

\item{bed.file}{a character string with with the bed file name or NULL if plink.prefix is given}

\item{out.prefix}{a character string with the prefix of the output file (will be named <out.prefix>.data)}

\item{no.web}{logical. If 'TRUE' PLINK will be called with noweb option}

\item{recode.pheno}{logical. If 'TRUE' phenotype coding will be changed from 1, 2 to 0, 1
for unaffecteds and affecteds, respectively}

\item{zip}{logical. If 'TRUE' ouput file will be zipped}
}
\description{
This function converts binary PLINK files to input files for Random Jungle. It uses PLINK
(\url{http://pngu.mgh.harvard.edu/~purcell/plink/}) to
change genotypes to 0, 1, 2 coding giving the number of minor alleles. This file is
further modified (e.g. the minor alleles are cut from the SNP names and the columns
with pedigree and gender information are removed).
}
\details{
Note:
This function assumes that the following programs are in the path:
\itemize{
  \item PLINK
  \item sed
  \item cut
  \item awk
  \item paste
  \item gzip
}
}

