% Generated by roxygen2 (4.0.0): do not edit by hand
\name{random.forest.runs}
\alias{random.forest.runs}
\title{Generate several random forests with different random number seeds using R.}
\usage{
random.forest.runs(data, out.prefix = NULL, type = "classification", ntree,
  mtry, nodesize, var.pheno, no.runs, seed = 123456, no.threads = 1,
  map.file = NULL)
}
\arguments{
\item{data}{data.frame with phenotype and predictor variables}

\item{out.prefix}{a character string with the prefix of the output files. The function generates two files: <out.prefix>_all_runs.info (errors and minimal importance scores in each run) and <out.prefix>_all_runs_relative.importance (relative variable importance for all variables in each run). If NULL, results will be returned as list.}

\item{no.runs}{number of forests to be generated}

\item{type}{type of random forest analysis (must be either 'classification' or 'regression')}

\item{ntree}{number of trees in each forest}

\item{mtry}{number of variables selected for each split}

\item{nodesize}{minimum size of terminal nodes. Note that it should be set to about 10\% of the sample size.}

\item{var.pheno}{name of column containing phenotype information}

\item{seed}{random number seed for first forest (additional forests will use <seed> + 1, <seed + 2>, ...)}

\item{no.threads}{number of threads for generating random forests in parallel mode}

\item{map.file}{a character string with name of map file (either MAP or BIM format)}
}
\value{
if out.prefix is not specified, a list with two elements:

info errors and minimal importance scores in each run

var.imp.rel: relative variable importance for all variables in each run
}
\description{
This function generates several random forests for the same data set using different
random number seeds. It uses the R package parallelRandomForest
\url{https://bitbucket.org/mkuhn/parallelrandomforest}.
}

