% Generated by roxygen2 (4.0.0): do not edit by hand
\name{random.jungle}
\alias{random.jungle}
\title{Generate random jungle.}
\usage{
random.jungle(data.file, out.prefix, type = "classification", ntree, mtry,
  nodesize, var.pheno = "PHENOTYPE", seed = 123456, no.threads = 1)
}
\arguments{
\item{data.file}{a character string with the name of the input file. This file might be
generated from binary PLINK files using \code{\link{prepare.input.file.rj}}.}

\item{out.prefix}{a character string with the prefix of the output files. For description
of the different files see manual for Random Jungle.}

\item{type}{type of random forest analysis (must be either 'classification' or 'regression')}

\item{ntree}{number of trees in the forest}

\item{mtry}{number of variables selected for each split}

\item{nodesize}{minimum size of terminal nodes. Note that it should be set to about 10\% of the sample size.}

\item{var.pheno}{name of column containing phenotype information (default value 'PHENOTYPE' used by \code{\link{prepare.input.file.rj}}).}

\item{seed}{random number seed}

\item{no.threads}{number of threads for running Random Jungle in parallel mode}
}
\description{
This function generates a random forest using the software Random Jungle
\url{http://imbs-luebeck.de/imbs/de/node/227}.
}
\details{
Note:
This function assumes that the program rjungle is in the path.
}

