% Generated by roxygen2 (4.0.0): do not edit by hand
\name{scatterplot.rel.imp.pvalue}
\alias{scatterplot.rel.imp.pvalue}
\title{Plot to compare results of random forest variable selection and statistical test.}
\usage{
scatterplot.rel.imp.pvalue(res.rf = NULL, res.test = NULL,
  rel.imp.file = NULL, pvalue.file = NULL, thresholds = c(2, 5 * 10^-8),
  var.highlight = NULL, main = "")
}
\arguments{
\item{res.rf}{data.frame with random forest results (columns 'varname', 'chr', 'pos', 'rel_imp_min') or NULL if results should be loaded from rel.imp.file}

\item{res.test}{data.frame with test results (columns 'SNP', 'CHR', 'BP', 'P') or NULL if results should be loaded from pvalue.file}

\item{rel.imp.file}{a character string with name of file with results from random forest variable selection (columns 'varname', 'chr', 'pos', 'rel_imp_min') or NULL if parameter res.rf is used.}

\item{pvalue.file}{a character string with name of file with PLINK results from test (columns 'SNP', 'CHR', 'BP', 'P') or NULL if parameter res.test is used.}

\item{thresholds}{vector with thresholds for relative importance and p-value (defaults: 2, 5*10^-8)}

\item{var.highlight}{vector with SNPs that will be plotted in red}

\item{main}{title for plot (default none)}
}
\description{
This function generates a scatterplot of minimal relative importance of random forest variable selection and log transformed p-values of a statistical test.
}

